% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalPval.R
\name{empiricalPval}
\alias{empiricalPval}
\title{empiricalPval}
\usage{
empiricalPval(observedStats, shuffledStats, step = 1)
}
\arguments{
\item{observedStats}{a vector of observed statistics}

\item{shuffledStats}{a vector of shuffled statistics}
}
\value{
a vector of p values taking values in [0,1]
}
\description{
empiricalPval estimates the single-sided empirical pvalues of observed 
statistics based on the shuffled statistics
}
\examples{

observedStats = qnorm( seq(0,1,0.1) )
shuffledStats = rnorm(100000)
# estimate the single-sided empirical pvalue
emp = getEmpiricalNum(observedStats,shuffledStats) 
cat(round(emp,2),'\\n') 

}
