% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fantom5.TF}
\alias{Fantom5.TF}
\title{Fantom5.TF}
\format{A list with two items:
\describe{
  \item{net}{a sparse matrix, the connectivity between terms and genes, comprising 500 Positional Weight Matrices (PWM) and 21964 genes}
  \item{annot}{a data frame, description of terms}
  ...
}}
\source{
\url{http://www.swissregulon.unibas.ch}
}
\usage{
Fantom5.TF
}
\description{
The human transcript promoter locations were obtained from Fantom5. 
Based on the promoter locations, the tool MotEvo was used to predict the human transcriptional factor (TF) target sites.
}
\keyword{datasets}
