% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSet2Net.R
\name{geneSet2Net}
\alias{geneSet2Net}
\title{geneSet2Net}
\usage{
geneSet2Net(term, geneset, value = NULL, sep = ",")
}
\arguments{
\item{term}{a vector of character values incidating the names of gene sets, like pathway names and miRNA names.}

\item{geneset}{a vector of character values, where each value is a gene list separated by 'sep'.}

\item{value}{a vector of numeric values indicating the connectivity of between terms and genes. It could take either discrete values (0 and 1) or continuous vlaues.}

\item{sep}{a character which separates the genes in the geneset.}
}
\value{
a matrix of numeric values where the column corresponds to the term and the row corresponds to the geneset.
}
\description{
geneSet2Net transforms gene sets to a matrix, which represents the connectivity between terms and genes.
}
\seealso{
\code{\link{geneSet2sparseMatrix}}; \code{\link{gmt2geneSet}};
}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
