% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIGSEA.R
\name{GIGSEA}
\alias{GIGSEA}
\title{GIGSEA}
\usage{
GIGSEA(data, geneCol, fcCol, weightCol = NULL,
  geneSet = c("MSigDB.KEGG.Pathway", "MSigDB.TF", "MSigDB.miRNA",
  "Fantom5.TF", "TargetScan.miRNA", "GO", "LINCS.CMap.drug"),
  permutationNum = 100, outputDir = getwd(), MGSEAthres = NULL)
}
\arguments{
\item{data}{a data frame comprising comlumns: gene names (characer), differential gene expression (numeric) and sample weights (numeric and optional)}

\item{geneCol}{an integer or a character indicating the column of gene name}

\item{fcCol}{an integer or a character indicating the column of differential gene expression}

\item{weightCol}{an integer or a character indicating the column of sample weights}

\item{geneSet}{a vector of characters indicating the gene sets of interest. It takes values from "MSigDB.KEGG.Pathway","MSigDB.TF","MSigDB.miRNA","Fantom5.TF","TargetScan.miRNA","GO" and "LINCS.CMap.drug"}

\item{permutationNum}{an integer indicating the number of permutation}

\item{outputDir}{a character value indicating the directory for saving the results}

\item{MGSEAthres}{an integer value indicating the thresfold for performing MGSEA. When the number of gene sets is smaller than MGSEAthres, we perform MGSEA.}
}
\description{
GIGSEA performs both SGSEA and MGSEA for a given list of gene sets, and write out the results.
}
\examples{

library(GIGSEA)
library(Matrix)
GIGSEA(data, geneCol='gene', fcCol='fc', weightCol= 'weights', geneSet=c("MSigDB.KEGG.Pathway","MSigDB.TF","MSigDB.miRNA","Fantom5.TF","TargetScan.miRNA","GO","LINCS.CMap.drug"), permutationNum=100, outputDir=getwd())
dir()


}
