% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationMultipleLm.R
\name{permutationMultipleLm}
\alias{permutationMultipleLm}
\title{permutationMultipleLm}
\usage{
permutationMultipleLm(fc, net, weights = rep(1, nrow(net)), num = 100)
}
\arguments{
\item{fc}{a vector of numeric values representing gene expression fold change}

\item{net}{a matrix of numeric values in the size of gene number x gene set number, representing the connectivity between genes and gene sets}

\item{weights}{a vector of numeric values representing the weights of permuated genes}

\item{num}{an integer value representing the number of permutations}
}
\value{
a data frame comprising the following columns:
\itemize{
\item {term} a vector of character incidating the names of gene sets.
\item {usedGenes} a vector of numeric values indicating the number of genes used in the model.
\item {Estimate} a vector of numeric values indicating the regression coefficients.
\item {Std..Error} a vector of numeric values indicating the standard errors of regression coefficients.
\item {t.value} a vector of numeric values indicating the t-statistics of regression coefficients.
\item {observedPval} a vector of numeric values [0,1] indicating the p values from the multiple weighted regression model.
\item {empiricalPval} a vector of numeric values [0,1] indicating the empirical p values from the permutation test.
}
}
\description{
permutationMultipleLm is a permutation test to calculate the empirical p values for a weighted multiple linear regression.
}
\seealso{
\code{\link{orderedIntersect}}; \code{\link{permutationMultipleLmMatrix}};
}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
