% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationSimpleLm.R
\name{permutationSimpleLm}
\alias{permutationSimpleLm}
\title{permutationSimpleLm}
\usage{
permutationSimpleLm(fc, net, weights = rep(1, nrow(net)), num = 100)
}
\arguments{
\item{fc}{a vector of numeric values representing the gene expression fold change}

\item{net}{a matrix of numeric values in the size of gene number x gene set number, representing the connectivity betweeen genes and gene sets}

\item{weights}{a vector of numeric values representing the weights of permuted genes}

\item{num}{a vector of integer values representing the number of permutations}
}
\value{
a data frame comprising the following columns:
\itemize{
\item {term} a vector of character values incidating the name of gene set.
\item {usedGenes} a vector of numeric values indicating the number of genes used in the model.
\item {Estimate} a vector of numeric values indicating the regression coefficients.
\item {Std..Error} a vector of numeric values indicating the standard errors of regression coefficients.
\item {t.value} a vector of numeric values indicating the t-statistics of regression coefficients.
\item {observedPval} a vector of numeric values [0,1] indicating the p values from weighted simple regression model.
\item {empiricalPval} a vector of numeric values [0,1] indicating the empirical p values from the permutation test.
}
}
\description{
permutationSimpleLm is a permutation test to calculate the empirical p values for a weighted simple linear regression.
}
\seealso{
\code{\link{orderedIntersect}}; \code{\link{permutationSimpleLmMatrix}};
}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
