% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TargetScan.miRNA}
\alias{TargetScan.miRNA}
\title{TargetScan.miRNA}
\format{A list with two items:
\describe{
  \item{net}{a sparse matrix, the connectivity between terms and genes, 
  comprising 87 miRNA seed motifs and 9861 genes}
  \item{annot}{a data frame, description of terms}
  ...
}}
\source{
\url{http://www.targetscan.org}
}
\usage{
TargetScan.miRNA
}
\description{
Gene sets of predicted human miRNA targets were obtained from TargetScan. 
TargetScan groups miRNAs that have identical subsequences at positions 2 
through 8 of the miRNA, i.e. the 2-7 seed region plus the 8th nucleotide, 
and provides predictions for each such seed motif.
}
\keyword{datasets}
