% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt2geneSet.R
\name{gmt2geneSet}
\alias{gmt2geneSet}
\title{gmt2geneSet}
\usage{
gmt2geneSet(gmt, termCol = 1, nonGeneCol = termCol, singleValue = NULL)
}
\arguments{
\item{gmt}{a vector of characters. Each item is a list of words comprising a term and its corresponding gene set, which are separated by tab.}

\item{termCol}{an integer value indicating in each item of gmt, which word is the term , by default, 1.}

\item{nonGeneCol}{an integer value indicating in each item of gmt, which words are not the gene set, by default, termCol.}

\item{singleValue}{a numeric value, which assigns the same value to all genes in a given gene set. This is useful when combining together the up-regulated gene sets (often time, singleValue=1) and the down-regulated gene sets (often time, singleValue=-1)}
}
\value{
a data frame, comprising three vectors: term (like pathway names), geneset (a gene symbol list separate by comma), and value (either discrete or continuous separated by comma)
}
\description{
gmt2geneSet transforms a gmt format file into geneSets.
}
\examples{

library(GIGSEA)
library(Matrix)

# downlaod the gmt file
gmt <- readLines('http://amp.pharm.mssm.edu/CREEDS/download/single_drug_perturbations-v1.0.gmt')

# obtain the index of up-regulated and down-regulated gene sets
index_up <- grep('-up',gmt)
index_down <- grep('-dn',gmt)

# transform the gmt file into gene sets. The gene set is a data frame, comprising three vectors: term (here is drug), geneset (a gene symbol list separate by comma), and value (1 and -1 separate by comma)
gff_up <- gmt2geneSet( gmt[index_up] , termCol=c(1,2) , singleValue = 1 )
gff_down <- gmt2geneSet( gmt[index_down] , termCol=c(1,2) , singleValue = -1 )

# as following, combine the up-regulated and down-regulated gene sets together, and use value of 1 and -1 to indicate their direction:
# extract the drug names
term_up <- sapply( gff_up$term , function(x) gsub('-up','',x) )
term_down <- sapply( gff_down$term , function(x) gsub('-dn','',x) )
all(term_up==term_down)

# combine the up-regulated and down-regulated gene names for each drug perturbation
geneset <- sapply( 1:nrow(gff_up) , function(i) paste(gff_up$geneset[i],gff_down$geneset[i],sep=',') )

# use 1 and -1 to indicate the direction of the up-regulated and down-regulated genes, respectively
value <- sapply( 1:nrow(gff_up) , function(i) paste(gff_up$value[i],gff_down$value[i],sep=',') )

# transform the gene set into matrix, where the row represents the gene, the column represents the drug perturbation, and each entry takes values of 1 and -1
net1 <- geneSet2Net( term=term_up , geneset=geneset , value=value )
# transform the gene set into sparse matrix, where the row represents the gene, the column represents the drug perturbation, and each entry takes values of 1 and -1
net2 <- geneSet2sparseMatrix( term=term_up , geneset=geneset , value=value )
tail(net1[,1:30])
tail(net2[,1:30])
# the size of sparse matrix is much smaller than the matrix
format( object.size(net1), units = "auto")
format( object.size(net2), units = "auto")

}
\references{

}
\seealso{
\code{\link{geneSet2Net}}; \code{\link{geneSet2sparseMatrix}};
}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
