% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationMultiLmMatrix.R
\name{permutationMultiLmMatrix}
\alias{permutationMultiLmMatrix}
\title{permutationMultiLmMatrix}
\usage{
permutationMultiLmMatrix(fc, net, weights = rep(1, nrow(net)), num = 100,
  step = 1000)
}
\arguments{
\item{fc}{}

\item{net}{}

\item{weights}{}

\item{num}{}

\item{step}{}
}
\description{
permutationMultiLmMatrix
}
\examples{

}
